+++ CHANGELOG.md
@@ -61,6 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Webhook and PubSub total count in the Console.
 - DevEUI is set when creating ABP devices via CLI.
 - CLI now shows all supported enum values for LoraWAN fields.
+- Application Server does not crash when retrieving a webhook template that does not exist if no template repository has been configured.
+- Application Server does not crash when listing webhook templates if no template repository has been configured.
 
 ### Security